feat(server): move files
